add method to load duration for time constraints
also move time constraint to separate module
